refactor frame drawing nodes and functions
