+++ mmpose/core/visualization/__init__.py
@@ -1,11 +1,13 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .effects import (apply_bugeye_effect, apply_firecracker_effect,
-                      apply_hat_effect, apply_sunglasses_effect)
+from .effects import apply_bugeye_effect, apply_sunglasses_effect
 from .image import (imshow_bboxes, imshow_keypoints, imshow_keypoints_3d,
                     imshow_mesh_3d)
 
 __all__ = [
-    'imshow_keypoints', 'imshow_keypoints_3d', 'imshow_bboxes',
-    'apply_bugeye_effect', 'apply_sunglasses_effect', 'imshow_mesh_3d',
-    'apply_hat_effect', 'apply_firecracker_effect'
+    'imshow_keypoints',
+    'imshow_keypoints_3d',
+    'imshow_bboxes',
+    'apply_bugeye_effect',
+    'apply_sunglasses_effect',
+    'imshow_mesh_3d',
 ]
