+++ tools/webcam/webcam_apis/nodes/helper_node.py
