+++ tools/webcam/webcam_apis/utils/__init__.py
@@ -5,10 +5,13 @@ from .message import FrameMessage, Message, VideoEndingMessage
 from .misc import (ImageCapture, copy_and_paste, expand_and_clamp,
                    get_cached_file_path, is_image_file, limit_max_fps,
                    load_image_from_disk_or_url, screen_matting)
+from .pose import (get_eye_keypoint_ids, get_face_keypoint_ids,
+                   get_wrist_keypoint_ids)
 
 __all__ = [
     'BufferManager', 'EventManager', 'FrameMessage', 'Message',
     'limit_max_fps', 'VideoEndingMessage', 'load_image_from_disk_or_url',
     'get_cached_file_path', 'screen_matting', 'expand_and_clamp',
-    'copy_and_paste', 'is_image_file', 'ImageCapture'
+    'copy_and_paste', 'is_image_file', 'ImageCapture', 'get_eye_keypoint_ids',
+    'get_face_keypoint_ids', 'get_wrist_keypoint_ids'
 ]
